% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggflower.R
\docType{data}
\name{geom_flower}
\alias{geom_flower}
\alias{GeomFlower}
\title{Create the flower geom}
\usage{
geom_flower(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = F,
  show.legend = NA,
  inherit.aes = T,
  shape = "normal",
  ...
)
}
\arguments{
\item{shape}{shape of the petal, one of "normal", "cos", "silo", "heart"}

\item{x}{categorical variable to group by}

\item{y}{the height of the group (y > 0)}
}
\description{
Create the flower geom
}
\examples{
n <- 8
dframe <- data.frame(index = 1:n, y = n:1)
dframe |>
  ggplot(aes(x = index, y = y)) +
  geom_flower(aes(fill = factor(index)))
dframe |>
  ggplot(aes(x = index, y = y)) +
  geom_flower(aes(fill = factor(index))) +
  coord_polar(theta = "x")
dframe |>
  ggplot(aes(x = index, y = y)) +
  geom_flower(aes(fill = factor(index)), shape = "heart") +
  coord_polar(theta = "x")
}
\keyword{datasets}
