% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggflower.R
\docType{data}
\name{geom_flower}
\alias{geom_flower}
\alias{GeomFlower}
\title{Create the flower geom}
\usage{
geom_flower(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = F,
  show.legend = NA,
  inherit.aes = T,
  disk.size = 1,
  disk.colour = "lightyellow",
  shape = "normal",
  ...
)
}
\arguments{
\item{shape}{shape of the petal, one of "normal", "cos", "circle", "heart"}

\item{x}{categorical variable to group by}

\item{y}{the height of the group (y > 0)}
}
\description{
Create the flower geom
}
\examples{
n <- 8
dframe <- data.frame(index = 1:n, y = n:1)
# scaling is not correct for cartesian coordinates
# squaring the y values ensures that heights are linear in y
dframe |>
  ggplot(aes(x = index, y = y^2)) +
  geom_flower(aes(fill = factor(index)))
# use polar coordinates
dframe |>
  ggplot(aes(x = index, y = y)) +
  geom_flower(aes(fill = factor(index)),
  shape = "cos", colour = NA, alpha = 0.8) +
  geom_flower(aes(x=1:8, y=2),linewidth=0.125,fill = NA, colour="lightyellow", shape = "cos") +
  geom_flower(aes(x=1:8, y=4),linewidth=0.25,fill = NA, colour="lightyellow", shape = "cos") +
  geom_flower(aes(x=1:8, y=6),linewidth=0.125,fill = NA, colour="lightyellow", shape = "cos") +
 # annotate("point", x = 1, y = -3, color="lightyellow", size=20) +
  geom_flower(aes(x=1:8, y=0),linewidth=0.25,fill = NA, colour="yellow", shape = "cos") +
  coord_polar() +
  theme_void() +
  theme(legend.position="bottom")
dframe |>
  ggplot(aes(x = index, y = y)) +
  geom_flower(aes(fill = factor(index)), shape = "circle") +
  coord_polar()
}
\keyword{datasets}
